make Key function more generically usable
