+++ cmds/ocm/commands/misccmds/rsakeypair/cmd.go
@@ -13,7 +13,6 @@ import (
 
 	parse "github.com/mandelsoft/spiff/dynaml/x509"
 	"github.com/mandelsoft/vfs/pkg/vfs"
-	utils2 "github.com/open-component-model/ocm/pkg/utils"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 
@@ -26,6 +25,7 @@ import (
 	"github.com/open-component-model/ocm/pkg/out"
 	"github.com/open-component-model/ocm/pkg/signing"
 	"github.com/open-component-model/ocm/pkg/signing/handlers/rsa"
+	utils2 "github.com/open-component-model/ocm/pkg/utils"
 )
 
 var (
