+++ cmds/ocm/commands/ocmcmds/common/options/signoption/option.go
@@ -8,10 +8,9 @@ import (
 	"crypto/x509"
 	"strings"
 
-	"github.com/open-component-model/ocm/cmds/ocm/commands/common/options/keyoption"
-	"github.com/open-component-model/ocm/pkg/utils"
 	"github.com/spf13/pflag"
 
+	"github.com/open-component-model/ocm/cmds/ocm/commands/common/options/keyoption"
 	"github.com/open-component-model/ocm/cmds/ocm/commands/ocmcmds/common/options/hashoption"
 	"github.com/open-component-model/ocm/cmds/ocm/pkg/options"
 	"github.com/open-component-model/ocm/pkg/contexts/clictx"
@@ -24,6 +23,7 @@ import (
 	"github.com/open-component-model/ocm/pkg/signing"
 	"github.com/open-component-model/ocm/pkg/signing/handlers/rsa"
 	"github.com/open-component-model/ocm/pkg/signing/hasher/sha256"
+	"github.com/open-component-model/ocm/pkg/utils"
 )
 
 func From(o options.OptionSetProvider) *Option {
