+++ cmds/ocm/topics/ocm/labels/topic.go
@@ -5,10 +5,10 @@
 package topicocmlabels
 
 import (
-	"github.com/open-component-model/ocm/pkg/contexts/ocm/valuemergehandler"
 	"github.com/spf13/cobra"
 
 	"github.com/open-component-model/ocm/pkg/contexts/clictx"
+	"github.com/open-component-model/ocm/pkg/contexts/ocm/valuemergehandler"
 )
 
 func New(ctx clictx.Context) *cobra.Command {
